ja/buscando os candidatos e provas antes de cadastrar

+++ backend/src/main/java/br/com/basis/sgp/servico/dto/AvaliacaoCadastroDTO.java
@@ -17,10 +17,10 @@ public class AvaliacaoCadastroDTO {
     private LocalDate data;
 
     @NotNull
-    private Integer idCandidato;
+    private Long idCandidato;
 
     @NotNull
-    private Integer idProva;
+    private Long idProva;
 
     private BigDecimal aproveitamento;
 }
